Mining | Complexity & Reward rules

+++ Microcoin/Blockchain/Mining/MiningBlockInfo.cs
@@ -13,7 +13,7 @@ namespace Microcoin.Blockchain.Mining
         public string PreviousBlockHash { get; set; } = "";
         public string MinerPublicKey { get; set; } = "";
         public decimal MinerReward { get; set; } = 0;
-        public double Complexity { get; set; } = 0;
+        public int Complexity { get; set; } = 0;
         public DateTime CreateTime { get; set; } = DateTime.UtcNow;
     }
 }
